🤖 Refactor(server): 테스트 코드 test 디렉토리로 이동
